Moved Covalence library out of Covalence directory
